chore(geovis): move makeMapMock to shared helpers directory
