+++ packages/geovis/tests/unit/tests/adapters/maplibre/legendFillPaint.test.ts
@@ -4,7 +4,7 @@ import {
   reapplyLegendDrivenFillPaint,
   setPaintWhenReady,
 } from 'src/adapters/maplibre/legendFillPaint';
-import { makeMapMock } from 'tests/unit/__mocks__/maplibre-gl';
+import { makeMapMock } from 'tests/unit/helpers/makeMapMock';
 
 jest.mock('maplibre-gl', () => {
   return {
